Move `service-config` into `config`
