+++ src/sidebar/config/test/service-config-test.js
@@ -1,7 +1,7 @@
 import serviceConfig from '../service-config';
 
-describe('serviceConfig', function () {
-  it('returns null if services is not an array', function () {
+describe('config/service-config', () => {
+  it('returns null if services is not an array', () => {
     const settings = {
       services: 'someString',
     };
@@ -9,7 +9,7 @@ describe('serviceConfig', function () {
     assert.isNull(serviceConfig(settings));
   });
 
-  it('returns null if the settings object has no services', function () {
+  it('returns null if the settings object has no services', () => {
     const settings = {
       services: [],
     };
@@ -17,7 +17,7 @@ describe('serviceConfig', function () {
     assert.isNull(serviceConfig(settings));
   });
 
-  it('returns the first service in the settings object', function () {
+  it('returns the first service in the settings object', () => {
     const settings = {
       services: [
         {
